component renamed
